\name{plot.exposure.levels}
\alias{plot.exposure.levels}
\title{Plotting the exposures to infection on vitamin D status curves.}
\description{
  Plot the exposures to infection as points on top of the vitamin D curves plotted from \code{plot.vitd.curve}.
  }
  \usage{
    \method{plot}{exposure.levels}( x, col = "blue", ... )
  }
  \arguments{
    \item{x}{An object of class \code{exposure.levels}.}
    \item{col}{Character, value for the colour of the points.}
    \item{\dots}{Additional arguments to be passed to the \code{plot} function.}
  }

\author{Rebecca Mangan and Jason Wyse}