\name{print.exposure.levels}
\alias{print.exposure.levels}
\title{Print summary of an exposure.levels object.}
\description{
  Print a summary of a exposure.levels object.
}
\usage{
  \method{print}{exposure.levels}( x, \dots )
}
\arguments{
  \item{x}{An object of class \code{exposure.levels}.}
   \item{\dots}{Optional arguments to lower level functions.}
}

\author{Rebecca Mangan and Jason Wyse}

