\name{print.infection.count}
\alias{print.infection.count}
\title{Print a summary of a infection.count object}
\description{
  Print a summary of a infection.count object.
}
\usage{
  \method{print}{infection.count}( x, \dots )
}
\arguments{
  \item{x}{An object of class \code{infection.count}.}
   \item{\dots}{Optional arguments to lower level functions.}
}

\author{Rebecca Mangan and Jason Wyse}
