\name{print.power.calc}
\alias{print.power.calc}
\title{Summary of a power.calc object}
\description{
  Print a summary of a power.calc object.
}
\usage{
  \method{print}{power.calc}( x, \dots )
}
\arguments{
  \item{x}{An object of class \code{power.calc}.}
   \item{\dots}{Optional arguments to lower level functions.}

}

\author{Rebecca Mangan and Jason Wyse}


