\name{print.vitd.curve}
\alias{print.vitd.curve}
\title{Print summary of a vitd.curve object}
\description{
  Print a summary of a vitd.curve object.
}
\usage{
  \method{print}{vitd.curve}( x, \dots )
}
\arguments{
  \item{x}{An object of class \code{vitd.curve}.}
   \item{\dots}{Optional arguments to lower level functions.}

}

\author{Rebecca Mangan and Jason Wyse}
