\name{summary.exposure.levels}
\alias{summary.exposure.levels}
\title{Summary of a vitd.curve object}
\description{
  Print a summary of a exposure.levels object.
}
\usage{
  \method{summary}{exposure.levels}( object, \dots )
}
\arguments{
  \item{object}{An object of class \code{exposure.levels}.}
   \item{\dots}{Optional arguments to lower level functions.}
}

\author{Rebecca Mangan and Jason Wyse}

