\name{summary.infection.count}
\alias{summary.infection.count}
\title{Summary of a infection.count object}
\description{
  Print a summary of a infection.count object.
}
\usage{
  \method{summary}{infection.count}( object, \dots )
}
\arguments{
  \item{object}{An object of class \code{infection.count}.}
   \item{\dots}{Optional arguments to lower level functions.}
}

\author{Rebecca Mangan and Jason Wyse}
