\name{summary.power.calc}
\alias{summary.power.calc}
\title{Summary of a power.calc object}
\description{
  Print a summary of a power.calc object.
}
\usage{
  \method{summary}{power.calc}( object, \dots )
}
\arguments{
  \item{object}{An object of class \code{power.calc}.}
   \item{\dots}{Optional arguments to lower level functions.}

}

\author{Rebecca Mangan and Jason Wyse}


