\name{summary.vitd.curve}
\alias{summary.vitd.curve}
\title{Summary of a vitd.curve object}
\description{
  Print a summary of a vitd.curve object.
}
\usage{
  \method{summary}{vitd.curve}( object, \dots )
}
\arguments{
  \item{object}{An object of class \code{vitd.curve}.}
   \item{\dots}{Optional arguments to lower level functions.}

}

\author{Rebecca Mangan and Jason Wyse}
