\name{plot.exposure.levels}
\alias{plot.exposure.levels}
\title{Plotting the exposures to infection on the vitamin D curves.}
\description{
  Plot the exposures to infection as points on top of the vitamin D curves plotted from \code{plot.vitd.curve}.
  }
  \usage{
    \method{plot}{exposure.levels}( x, pch = 1, cex = 1.3, col = "red", ... )
  }
  \arguments{
    \item{x}{An object of class \code{exposure.levels}.}
    \item{pch}{Integer, specifying symbol for the points.}
    \item{cex}{Numeric, value giving the amount by which plotting the points should be magnified.}
    \item{col}{Character, value for the colour of the points.}
    \item{\dots}{Additional arguments to be passed to the \code{plot} function.}
  }

\author{Rebecca Mangan}