\name{plot.vitd.curve}
\alias{plot.vitd.curve}
\title{Plotting a vitd.curve object}
\description{
  Plot the vitamin D curves for a group of size N.
  }
\usage{
  \method{plot}{vitd.curve}( x, main = " ", xlab = " ", ylab = "25 Hydroxy Vitamin D", \dots )
}
\arguments{
  \item{x}{An object of class \code{vitd.curve}.}
  \item{main}{Main title for the plot.}
  \item{xlab}{A title for the x-axis.}
  \item{ylab}{A title for the y-axis.}
  \item{\dots}{Additional arguments to the \code{plot} function.}
}
